+++ tomcat/build.gradle
@@ -17,7 +17,6 @@ dependencies {
     implementation 'org.reflections:reflections:0.10.2'
     implementation 'ch.qos.logback:logback-classic:1.5.18'
     implementation 'org.apache.commons:commons-lang3:3.18.0'
-    implementation 'com.fasterxml.jackson.core:jackson-databind:2.16.0'
 
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher:1.13.3'
