+++ tomcat/src/main/java/org/apache/coyote/http11/common/ContentType.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.common;
 
 public enum ContentType {
     NONE(""),
